requirements.txt
python-dotenv
ipykernel  # Needed for notebooks
ipython  # needed for ipython
pandas  # data structures and operations for manipulating numerical tables
numpy<2  # adding support for large, multi-dimensional arrays and matrices. Restricting version to <2 for compatibility.
termcolor  # ANSI color formatting for output in terminal
tensorflow  # open source software library for high performance numerical computation
scikit-learn  # a free software machine learning library for the Python programming language.
matplotlib  # plotting library.
seaborn  # Python data visualization library based on matplotlib.
k-means-constrained  # Variant to K-means algorithm that enables to have clusters of minimum sample-size
statsmodels  # Tool for statistical calculations
jupyter  # Enable to open jupyter notebook
hdbscan  # Hierarchical Density-Based Spatial Clustering of Applications with Noise
umap-learn